fix tests

+++ tests/PageTest.php
@@ -10,18 +10,28 @@ class PageTest extends BLBaseTest
     public function tearDown()
     {
         parent::tearDown();
-        $this->page->forceDelete();
+        if($this->page) $this->page->forceDelete();
     }
 
     public function testCreateUser(){
         $hash = substr(md5(microtime()),0,5);
-        $email = 'test_user_'.$hash.'@test.com';
-        $name = 'test_user_'.$hash;
+        $title = 'test_page_'.$hash;
 
-        $this->page = $this->createPage();
+        $this->page = $this->createPage($title);
         $this->assertEquals('Model_Page',get_class($this->page));
-//        $this->assertEquals($this->page['email'],$email);
-//        $this->assertEquals($this->page['name'],$name);
+        $this->assertEquals($this->page['title'],$title);
+    }
+
+    public function testCreateUserFailType(){
+        try{
+            $this->page = $this->createPage(null,'top','shot');
+        }catch (Exception_InvalidField $e){
+            $this->assertInstanceOf('Exception_InvalidField',$e);
+            return;
+        }catch(Exception $e){
+            $this->fail('Incorrect Exception was thrown');
+        }
+        $this->fail('Exception_InvalidField was not thrown');
     }
 }
